+++ RobotCode/lib/miam_utils_arduino/PointTurn.cpp
@@ -10,7 +10,7 @@ namespace miam{
     namespace trajectory{
         PointTurn::PointTurn(TrajectoryConfig const& config,
                              RobotPosition const& startPoint,
-                             double const& endAngle):
+                             float const& endAngle):
             Trajectory(config),
             endAngle_(endAngle)
         {
@@ -18,7 +18,7 @@ namespace miam{
             make(startPoint);
         }
 
-        TrajectoryPoint PointTurn::getCurrentPoint(double const& currentTime)
+        TrajectoryPoint PointTurn::getCurrentPoint(float const& currentTime)
         {
             TrajectoryPoint output;
             output.position = startPoint_;
@@ -35,13 +35,13 @@ namespace miam{
             startPoint_ = startPoint;
             motionSign_ = 1.0;
             // Create trapezoid.
-            double length = moduloTwoPi(endAngle_ - startPoint.theta);
+            float length = moduloTwoPi(endAngle_ - startPoint.theta);
             if(length < 0)
                 motionSign_ = -1.0;
 
             // Compute max angular velocity and acceleration, taking into account wheel spacing.
-            double maxRobotAngularVelocity = config_.maxWheelVelocity / config_.robotWheelSpacing;
-            double maxRobotAngularAcceleration = config_.maxWheelAcceleration / config_.robotWheelSpacing;
+            float maxRobotAngularVelocity = config_.maxWheelVelocity / config_.robotWheelSpacing;
+            float maxRobotAngularAcceleration = config_.maxWheelAcceleration / config_.robotWheelSpacing;
 
             trapezoid_ = Trapezoid(length, 0.0, 0.0, maxRobotAngularVelocity, maxRobotAngularAcceleration);
             duration_ = trapezoid_.getDuration();
@@ -49,7 +49,7 @@ namespace miam{
         }
 
 
-        void PointTurn::replanify(double const& replanificationTime)
+        void PointTurn::replanify(float const& replanificationTime)
         {
             RobotPosition startPoint = getCurrentPoint(replanificationTime).position;
             make(startPoint);
