+++ RobotCode/lib/miam_utils_arduino/StraightLine.h
@@ -27,27 +27,27 @@
                     StraightLine(TrajectoryConfig const& config,
                                  RobotPosition const& startPoint,
                                  RobotPosition const& endPoint,
-                                 double const& startVelocity = 0.0,
-                                 double const& endVelocity = 0.0,
+                                 float const& startVelocity = 0.0,
+                                 float const& endVelocity = 0.0,
                                  bool const& backward = false);
 
-                    TrajectoryPoint getCurrentPoint(double const& currentTime);
+                    TrajectoryPoint getCurrentPoint(float const& currentTime);
 
-                    void replanify(double const& replanificationTime);
+                    void replanify(float const& replanificationTime);
 
                     /// \brief Return line angle.
                     ///
                     /// \return Line angle
-                    double getAngle();
+                    float getAngle();
                 private:
-                    void make(RobotPosition const& startPoint, double const& startVelocity); ///< Build (or rebuild) the trajectory.
+                    void make(RobotPosition const& startPoint, float const& startVelocity); ///< Build (or rebuild) the trajectory.
 
                     RobotPosition startPoint_; ///< Point where the trajectory started.
                     int motionSign_; ///< 1 or -1, to indicate direction of motion.
                     Trapezoid trapezoid_; ///< Velocity trapezoid.
 
                     RobotPosition endPoint_; ///< End position.
-                    double endVelocity_; ///< End velocity.
+                    float endVelocity_; ///< End velocity.
                     bool backward_;     ///< True if going backward.
             };
         }
